Rename: Recoil atoms 폴더 정리, 파일명 수정
