+++ src/hooks/useTheme.ts
@@ -1,6 +1,6 @@
 import { useCallback, useLayoutEffect } from 'react';
 import { useRecoilState } from 'recoil';
-import { themeState } from 'recoil/atoms/theme';
+import { themeState } from 'recoil/atoms/themeAtoms';
 
 function useTheme() {
   const [theme, setTheme] = useRecoilState(themeState);
